Initial Commit
This provides two templates, one for setting up Views for all
default FOLIO tables, and the other for setting up a Role that
has permissions to use these Views.  Note that USAGE must be granted
on the _mod_inventory_storage Schema, as well, in order to access
necessary Functions.

A Python script is provided to make use of these templates.  Provide
commandline switches to render the template for a particular tenant.

Options currently include:
 - tenant: the id of the tenant, which is assumed to prefix the module schema
 - schema: name for the reporting schema to be created for the Views
 - user: name of the Role to create and grant privileges for
 - pw: user password; if included, user Role is created, otherwise it is
assumed to already be present

To install, an administrator will need the rendered output from this
script, an SQL file which can then be loaded into the FOLIO database